Fix SandboxClassLoader to properly inherit from erstwhileClassLoader
The ClassLoader interface requires new ClassLoaders to inherit from old ClassLoader via the parent hierarchy. Before this change, it would inherit from the parent of erstwhileClassLoader, but still load classes from the original erstwhileClassLoader. This violates the ClassLoader interface, which requires all classes to be loaded in a singular chain.

Additionally, we first have to check if we already found the class previously in the ClassLoader. Then, we check if we should acquire the Class in question. If not, we delegate to our parent ClassLoader. This encovered another bug, in which on JDK >= 9, the SandboxClassLoader was incorrectly redefining classes from jdk.internal packages.

PiperOrigin-RevId: 249624247